third_party/blink/web_tests/flag-specific/highdpi/virtual/view-transition/external/wpt/css/css-view-transitions/parsing/view-transition-name-invalid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['view-transition-name'] = "auto" should not set the property value
  assert_equals: expected "" but got "auto"
Harness: the test ran to completion.
